refactor(core): use standard dom export
