+++ internal/design/ui/menus.md
@@ -16,7 +16,7 @@ Applications need lightweight, independent menus. Player settings also need opti
 - Keep each menu's committed open state, items, focus, and positioning in a framework-neutral core. Open and close interactions request changes; adapters commit resolved controlled or uncontrolled state.
 - Use compound parts rather than one configuration-heavy component so consumers can compose custom content while preserving behavior.
 - Keep base Menu independent of parent detection, navigation stacks, transition coordination, and player-setting selectors.
-- Bind nested settings explicitly through `@videojs/core/dom/menu-transition`, the existing React `Menu.TransitionRoot` and `Menu.TransitionView` parts, or `@videojs/html/ui/menu-transition`. The binding derives the active view from child menu roots and owns forward/back policy.
+- Bind nested settings explicitly through the shared core DOM coordinator, the existing React `Menu.TransitionRoot` and `Menu.TransitionView` parts, or `@videojs/html/ui/menu-transition`. The binding derives the active view from child menu roots and owns forward/back policy.
 - Generate the root panel inside the transition binding. Consumers wrap child roots only; they do not author a root-view part.
 - Use an ordinary menu item for a back row. Selecting it closes the child menu that owns it.
 - Keep player-setting labels separate. React presets render labels from option hooks; HTML registers `<media-menu-item-value>` through `@videojs/html/ui/menu-settings`.
@@ -32,7 +32,7 @@ Base core and HTML imports exclude nested navigation, transition measurement, an
 ## Current sources of truth
 
 - Base core and tests: `packages/core/src/core/ui/menu/` and `packages/core/src/dom/ui/menu/`
-- Optional shared coordinator: `packages/core/src/dom/ui/menu-transition/`
+- Shared DOM coordinator: `packages/core/src/dom/ui/menu/create-menu-transition.ts`
 - HTML base and optional bindings: `packages/html/src/ui/menu/` and `packages/html/src/define/ui/menu-*.ts`
 - React Menu namespace and transition binding: `packages/react/src/ui/menu/`
 - Detailed rationale and migration: `rfc/menu-view-transitions.md`
