--- packages/core/src/dom/menu-transition.ts
@@ -1,1 +1,0 @@
-export * from './ui/menu-transition/create-menu-transition';
